Add files via upload
Just remembered I'm supposed to be showing my work through progressive commits, so uploading the skeleton I threw together to start work with.

One core philosophy to the page object model is actually to write your tests with essentially pseudocode, then build the backend to make it work.  This keeps you from overengineering and also encourages the framework to develop in a way that makes sense to the product.  So I want my "happy path" login test to be as readable and simple as possible:
1) Go to login
2) Login (with default credentials- from env variables per requirements)
3) Assert that it brought me to the home page.